Add Anonymous Admins handling and logging to @admin_only
Anonymous Admins came with Bot API 5.0:

https://core.telegram.org/bots/api#november-4-2020

> Added the field sender_chat to the class Message, containing the
sender of a message which is a chat (group or channel). For backward
compatibility in non-channel chats, the field from in such messages will
contain the user 777000 for messages automatically forwarded to the
discussion group and the user 1087968824 (@GroupAnonymousBot) for
messages from anonymous group administrators.